Allow importing an existing key from the onboarding profile step
The key-import option only existed on the WelcomeSetup screen, which
renders solely when no workspace is configured. A returning member
whose workspace is already set up (new laptop with synced config, dev
worktree, re-onboarding) lands directly on the profile step and never
sees a way to reuse their key.

Extract the paste-or-drop nsec form out of WelcomeSetup into a shared
NostrKeyImportForm and add an "I already have a key" link on the
profile step that opens it. Import reuses the membership-denied swap
path: the identity changes, the flow remounts keyed on pubkey, and the
new key's relay profile reseeds the steps — a key that already
completed onboarding on this machine skips straight into the app.

Adds an e2e test for the workspace-already-set-up import scenario;
shared-form testids renamed welcome-nostr-* -> nostr-import-*.
