+++ desktop/src/features/onboarding/ui/OnboardingFlow.tsx
@@ -19,7 +19,11 @@ import { ONBOARDING_DEFAULT_THEME_NAME } from "@/shared/theme/theme-loader";
 import { StepProgress } from "@/shared/ui/step-progress";
 import { AvatarStep } from "./AvatarStep";
 import { MembershipDenied } from "./MembershipDenied";
-import type { OnboardingTransitionDirection } from "./OnboardingSlideTransition";
+import { NostrKeyImportForm } from "./NostrKeyImportForm";
+import {
+  type OnboardingTransitionDirection,
+  OnboardingSlideTransition,
+} from "./OnboardingSlideTransition";
 import { ProfileStep } from "./ProfileStep";
 import { SetupStep } from "./SetupStep";
 import { ThemeStep, preloadThemePreviewVars } from "./ThemeStep";
@@ -228,6 +232,11 @@ export function OnboardingFlow({
     setCurrentPage("profile");
   }, []);
 
+  const showKeyImportPage = React.useCallback(() => {
+    setTransitionDirection("forward");
+    setCurrentPage("key-import");
+  }, []);
+
   const saveProfileAndContinue = React.useCallback(
     async (nextPage: OnboardingPage) => {
       if (isProfileAdvancePending) {
@@ -365,7 +374,11 @@ export function OnboardingFlow({
       : profileStepState.saveRecovery,
   };
 
-  const importDeniedKey = React.useCallback(
+  // Swapping the identity changes the pubkey, which remounts this flow
+  // (keyed on pubkey in App.tsx) and re-runs the onboarding gate: the new
+  // key's relay profile reseeds the steps, and a key that already finished
+  // onboarding on this machine skips straight into the app.
+  const importExistingKey = React.useCallback(
     async (nsec: string) => {
       const identity = await importIdentity(nsec);
       relayClient.disconnect();
@@ -390,7 +403,7 @@ export function OnboardingFlow({
               }
             : undefined
         }
-        onImportKey={canBackToWorkspaceSetup ? undefined : importDeniedKey}
+        onImportKey={canBackToWorkspaceSetup ? undefined : importExistingKey}
         onRetry={() => {
           void saveProfileAndContinue(membershipRetryPage);
         }}
@@ -402,7 +415,9 @@ export function OnboardingFlow({
   return (
     <div
       className={`sprout-startup-shell flex items-center justify-center bg-background px-4 py-8 text-foreground ${
-        currentPage === "profile" || currentPage === "avatar"
+        currentPage === "profile" ||
+        currentPage === "avatar" ||
+        currentPage === "key-import"
           ? "sprout-onboarding-neutral-theme"
           : ""
       }`}
@@ -429,7 +444,7 @@ export function OnboardingFlow({
           }`}
           completeSegmentClassName="bg-primary/35"
           currentStep={
-            currentPage === "profile"
+            currentPage === "profile" || currentPage === "key-import"
               ? 2
               : currentPage === "avatar"
                 ? 3
@@ -451,6 +466,7 @@ export function OnboardingFlow({
                   }
                 : undefined,
               clearAvatarDraft: resetAvatarDraft,
+              importExistingKey: showKeyImportPage,
               onUploadingChange: setIsUploadingAvatar,
               skipForNow,
               submit: () => {
@@ -462,6 +478,28 @@ export function OnboardingFlow({
             direction={transitionDirection}
             state={profileStepState}
           />
+        ) : currentPage === "key-import" ? (
+          <OnboardingSlideTransition
+            className="flex w-full flex-col items-center text-center"
+            direction={transitionDirection}
+            transitionKey={`key-import-${transitionDirection}`}
+          >
+            <div className="w-full max-w-[440px]">
+              <h1 className="text-3xl font-semibold tracking-tight">
+                Use your existing key
+              </h1>
+              <p className="mt-3 text-sm leading-6 text-muted-foreground">
+                Import your Nostr private key to use that identity with Sprout.
+                If this key already has a profile on the relay, your name and
+                avatar are restored automatically.
+              </p>
+            </div>
+
+            <NostrKeyImportForm
+              onBack={showProfilePage}
+              onImport={importExistingKey}
+            />
+          </OnboardingSlideTransition>
         ) : currentPage === "avatar" ? (
           <AvatarStep
             actions={{
